% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_Logit.R
\name{Bayes_Logit}
\alias{Bayes_Logit}
\title{Bayesian Logistic Regression using Polya-Gamma data augmentation}
\usage{
Bayes_Logit(
  y,
  Xmat,
  m0 = rep(0, ncol(Xmat)),
  P0 = matrix(0, nrow = ncol(Xmat), ncol = ncol(Xmat)),
  numReps,
  thin,
  burninPerc,
  n_chains,
  start_mat
)
}
\arguments{
\item{y}{an n-length binary vector of outcomes}

\item{Xmat}{an n by p matrix of covariates}

\item{m0}{prior mean vector}

\item{P0}{prior variance matrix}

\item{numReps}{numeric total number of iterations}

\item{thin}{numeric number of iterations to take before each sample}

\item{burninPerc}{numeric proportion of numReps to toss as burn-in}

\item{n_chains}{numeric number of chains to initialize, if start_list is null}

\item{start_mat}{a matrix of start values, with each column corresponding to a chain. If null, n_chains sets number of chains}
}
\value{
a list with outputs
}
\description{
Bayesian Logistic Regression using Polya-Gamma data augmentation
}
