% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_SCR_logit.R
\name{Bayes_SCR_logit_eps}
\alias{Bayes_SCR_logit_eps}
\title{Bayesian Illness-Death Model with Weibull Baselines, Semi-Markov structure and logit model for immediate event}
\usage{
Bayes_SCR_logit_eps(
  y1,
  y_sm,
  delta1,
  delta1D,
  delta2,
  eps,
  vars1,
  vars2,
  vars3,
  varsD,
  data,
  na.action = "na.fail",
  subset = NULL,
  hyperParams,
  mcmcParams,
  n_chains,
  start_mat = NULL,
  frailty = TRUE,
  frail_path = NULL,
  logLHi_path = NULL
)
}
\arguments{
\item{y1}{time to first event (nonterminal, terminal only, censored)}

\item{y_sm}{sojourn time from end of terminal window (y1+eps) (not length n! Just as long as sum(delta1D))}

\item{delta1}{indicator for nonterminal event}

\item{delta1D}{indicator for nonterminal event followed by immediate death}

\item{delta2}{indicator for terminal event}

\item{eps}{width of "immediate" event window}

\item{vars1, vars2, vars3, varsD}{string vectors with variable names for each submodel}

\item{data}{data}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows. See \code{\link[stats]{model.frame}}.}

\item{hyperParams}{list of hyperparameters}

\item{mcmcParams}{list of mcmc control parameters}

\item{n_chains}{integer for number of chains to fit}

\item{start_mat}{a matrix of start values, with each column corresponding to a chain. If null, n_chains sets number of chains}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included.}

\item{frail_path}{filepath for where to write frailty output}

\item{logLHi_path}{filepath for where to write individual likelihood contribution output}
}
\value{
a list with outputs
}
\description{
Bayesian Illness-Death Model with Weibull Baselines, Semi-Markov structure and logit model for immediate event
}
