% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk_logitbayes.R
\name{calc_risk_logitbayes}
\alias{calc_risk_logitbayes}
\title{Calculate absolute risk profiles}
\usage{
calc_risk_logitbayes(
  para,
  Xmat1,
  Xmat2,
  Xmat3,
  XmatD,
  D_eps,
  t_cutoff,
  tol = 0.001,
  hazard,
  frailty = TRUE,
  beta2frail = 1,
  beta3frail = 1,
  betaDfrail = 1,
  type = "marginal",
  gamma = 1,
  model = "semi-markov",
  h3_tv = "none",
  h3tv_knots = NULL,
  logit_ind = TRUE,
  logit_tv = "none",
  logit_tv_knots = NULL,
  n_quad = 15
)
}
\arguments{
\item{para}{A numeric vector of parameters, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.}

\item{Xmat1, Xmat2, Xmat3}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{XmatD}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{D_eps}{Numeric scalar width of the "immediate event" window.}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}

\item{tol}{Numeric value for the tolerance of the numerical integration procedure.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to TRUE.}

\item{beta2frail, beta3frail, betaDfrail}{coefficients on the log-frailty in each hazard, only used for conditional prediction}

\item{type}{String either indicating 'marginal' for population-averaged probabilities,
or 'conditional' for probabilities computed at the specified gamma}

\item{gamma}{Numeric value indicating the fixed level of the frailty assumed for predicted probabilities,
if 'type' is set to 'conditional'}

\item{model}{String specifying the transition assumption}

\item{h3_tv}{String indicating whether there is an effect of t1 on hazard 3.}

\item{h3tv_knots}{for piecewise effect of t1 in h3, these are the knots at which the effect jumps}

\item{logit_ind}{Boolean for whether to include logit submodel.}

\item{logit_tv}{String indicating whether there is an effect of t1 on the logit model.}

\item{logit_tv_knots}{for piecewise effect of t1 in h3, these are the knots at which the effect jumps}

\item{n_quad}{Scalar for number of Gaussian quadrature points used to evaluate numerical integral of B-spline.}
}
\value{
if Xmat has only one row, and t_cutoff is a scalar, then returns a 4 element row matrix
  of probabilities. If Xmat has \code{n} rows, then returns an \code{n} by 4 matrix of probabilities.
  If Xmat has \code{n} rows and t_cutoff is a vector of length \code{s}, then returns an \code{s} by 4 by \code{n} array.
}
\description{
This function calculates absolute risk profiles under weibull baseline hazard specifications.
}
