% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghquad_pointsweights.R
\name{get_ghquad_pointsweights}
\alias{get_ghquad_pointsweights}
\title{Function to return quadrature points and weights}
\usage{
get_ghquad_pointsweights(n_quad)
}
\arguments{
\item{n_quad}{number of quadrature points}
}
\value{
A list with two named elements (\code{points} and \code{weights}), each
  of which is a numeric vector with length equal to the number of
  quadrature nodes
}
\description{
This function stores quadrature points used for numerical integration by
Gauss-Hermite rule. The values are pre-computed and stored for
number of nodes equal to 7, 11, 15, 21, 31, 41, 51, or 61.
}
