% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_metrics.R
\name{get_ipcw_mat}
\alias{get_ipcw_mat}
\title{Get inverse probability of censoring weights}
\usage{
get_ipcw_mat(y2, delta2, t_cutoff)
}
\arguments{
\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}
}
\value{
a vector.
}
\description{
This function returns a vector of inverse probability of censoring weights from an unadjusted Cox model
  for censoring times.
}
