% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_metrics.R
\name{get_ipcw_mat_logit}
\alias{get_ipcw_mat_logit}
\title{Get inverse probability of censoring weights}
\usage{
get_ipcw_mat_logit(y2, delta1, delta2, deltaD, t_cutoff)
}
\arguments{
\item{delta1, delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{deltaD}{indicator that is 1 if subject has delta1==1 and observed immediate event, 0 else}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}
}
\value{
a vector.
}
\description{
This function returns a vector of inverse probability of censoring weights from an unadjusted Cox model
  for censoring times.
}
