% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict_logit}
\alias{predict_logit}
\title{function to predict probabilities from bayesian logistic regression}
\usage{
predict_logit(object, xnew = NULL, alpha = 0.05, ...)
}
\arguments{
\item{object}{object fit by Bayes_Logit}

\item{xnew}{new matrix of values at which to predict}

\item{alpha}{significance level for credible interval}

\item{...}{extras}
}
\value{
a matrix
}
\description{
function predicts probabilitites
}
