% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict_term}
\alias{predict_term}
\title{dumb little function to predict outcomes given non-terminal event timing.}
\usage{
predict_term(
  object,
  x3new = NULL,
  xDnew = NULL,
  logit_ind = TRUE,
  D_eps = 0,
  tseq,
  type = "conditional",
  out = "long",
  alpha = 0.05,
  n_quad = 15,
  ...
)
}
\arguments{
\item{object}{object fit by Bayes_Logit}

\item{x3new, xDnew}{new matrix of values at which to predict}

\item{logit_ind}{augmented model or standard model}

\item{tseq}{sequence of values at which to make predictions}

\item{type}{conditional or marginal}

\item{out}{format for output}

\item{alpha}{significance level for credible interval}

\item{n_quad}{number of quadrature points for prediction}

\item{...}{extras}
}
\value{
a matrix
}
\description{
function predicts probabilitites
}
