% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simID_logit.R
\name{simID_logit}
\alias{simID_logit}
\title{The function that simulates independent/cluster-correlated semi-competing
  risks data under semi-Markov Weibull models with logit submodel for immediate terminal event.}
\usage{
simID_logit(
  x1,
  x2,
  x3,
  xD = NULL,
  beta1.true,
  beta2.true,
  beta3.true,
  alpha1.true,
  alpha2.true,
  alpha3.true,
  kappa1.true,
  kappa2.true,
  kappa3.true,
  theta.true,
  frailty_type = "gamma",
  beta2frail.true = 1,
  beta3frail.true = 1,
  betaDfrail.true = 1,
  beta3tv.true = NULL,
  h3tv_degree = 3,
  anyD = TRUE,
  betaD.true = NULL,
  betaDtv.true = NULL,
  Dtv_degree = 3,
  cens = c(0, 0)
)
}
\arguments{
\item{x1, x2, x3, xD}{Covariate matrices with \code{n} rows.}

\item{beta1.true, beta2.true, beta3.true, betaD.true}{Vectors of true regression parameter values.
The length of each vector should equal the number of columns in the corresponding covariate matrix.}

\item{alpha1.true, alpha2.true, alpha3.true, kappa1.true, kappa2.true, kappa3.true}{Vectors of true baseline parameter values.}

\item{theta.true}{True value for \eqn{\theta}.}

\item{frailty_type}{string denoting "gamma" for gamma-distributed frailty with variance \code{theta}, or "lognormal" for lognormal distributed frailty with log-frailty variance \code{theta}}

\item{beta2frail.true, beta3frail.true, betaDfrail.true}{scalar for the coefficient of the shared frailty in the logit submodel.}

\item{beta3tv.true, betaDtv.true}{Vectors of true regression parameter values for effects of T1 on the logit and h3 submodels.}

\item{h3tv_degree, Dtv_degree}{either the string "cs" indicating restricted cubic spline, or an integer for degree of time-varying hazard/odds ratio B-spline basis. (0 is piecewise constant)}

\item{anyD}{Boolean for whether to allow any "immediate" terminal events}

\item{cens}{A numeric vector of two elements. The right censoring times are generated from Uniform(\eqn{cens[1]}, \eqn{cens[2]}).}
}
\value{
returns a data.frame containing semi-competing risks outcomes from \code{n} subjects.
  It is of dimension \eqn{n\times 4}: the columns correspond to \eqn{y_1}, \eqn{\delta_1}, \eqn{y_2}, \eqn{\delta_2}. \cr
  \itemize{
  \item{y1}{a vector of \code{n} times to the non-terminal event}
  \item{y2}{a vector of \code{n} times to the terminal event}
  \item{delta1}{a vector of \code{n} censoring indicators for the non-terminal event time (1=event occurred, 0=censored)}
  \item{delta2}{a vector of \code{n} censoring indicators for the terminal event time (1=event occurred, 0=censored)}
  \item{deltaD}{a vector of \code{n} indicators whether the terminal event occurred immediately}
  }
}
\description{
The function that simulates independent/cluster-correlated semi-competing
  risks data under semi-Markov Weibull models with logit submodel for immediate terminal event.
}
